20/11

+++ Events/Client/messageCreate.js
@@ -3,10 +3,14 @@ module.exports = {
     async execute(message) {
         if (message.author.bot) return;
 
-        if (message.content.includes("https://") || message.content.includes("http://") || message.content.includes("discord.gg")) {
-            message.delete();
+        // Kiểm tra xem người gửi có vai trò phù hợp không
+        const allowedRoleIDs = ["1121667704298946600", "1121764644403560469"]; // Thay thế bằng ID của các vai trò được phép
+        const senderRoleIDs = message.member.roles.cache.map(role => role.id);
 
-            message.channel.send({content: `Này ${message.author}, không được gửi link ở đây nha!`})
+        if (!(allowedRoleIDs.some(roleID => senderRoleIDs.includes(roleID))) &&
+            (message.content.includes("https://") || message.content.includes("http://") || message.content.includes("discord.gg"))) {
+            message.delete();
+            message.channel.send({ content: `Này ${message.author}, bạn không được phép gửi link ở đây!` });
         }
     }
-}
+};
